test: split capability service composition tests by concern
